requirements.txt
sqlalchemy
pymysql
alembic
python-dotenv
fastapi
uvicorn
gunicorn
psycopg2-binary